message update

+++ src/app/new-researches/new-researches.component.ts
@@ -3,6 +3,9 @@ import { Title } from '@angular/platform-browser';
 import { data } from '../interfaces/revision.interface';
 import { GetDataService } from '../services/get-data.service';
 
+import Swal from 'sweetalert2'
+
+
 @Component({
   selector: 'app-new-researches',
   templateUrl: './new-researches.component.html',
@@ -11,16 +14,22 @@ import { GetDataService } from '../services/get-data.service';
 export class NewResearchesComponent implements OnInit {
 
   revisionData:data[]=[]
+
+  loading:Boolean=false;
   
   constructor(private title:Title, private getserviceData:GetDataService) { }
 
   ngOnInit(): void {
     this.title.setTitle("new researches");
 
+    this.loading=true 
+
     this.getserviceData.getRevisionData().subscribe(data=>{
       this.revisionData=data
     })
     
+    this.loading=false 
+
   }
 
   GeneralInformationText1:string=""
@@ -43,8 +52,51 @@ export class NewResearchesComponent implements OnInit {
     this.abstract=abst;
   }
 
+
+
+
+  
   deleteItem(item:any){
-    this.getserviceData.delete_from_NewResearch(item.id)
-    window.location.reload()
-  }
+  // ------------------------------------ code for design --------------------------------
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger',
+    },
+    // buttonsStyling: false,
+  })
+  
+  swalWithBootstrapButtons.fire({
+    title: 'Are you sure?',
+    text: "You want to remove this item!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: 'Yes, delete it!',
+    cancelButtonText: 'No, cancel! ',
+    reverseButtons: true,
+    
+  }).then((result) => {
+    if (result.isConfirmed) {
+      swalWithBootstrapButtons.fire(
+        'Deleted!',
+        'Your file has been deleted.',
+        'success'
+      ).then(()=>{
+        this.getserviceData.delete_from_NewResearch(item.id)
+        window.location.reload()
+      })
+      
+  }else if (
+      /* Read more about handling dismissals below */
+      result.dismiss === Swal.DismissReason.cancel
+    ) {
+      swalWithBootstrapButtons.fire(
+        'Cancelled',
+        'Your imaginary file is safe :)',
+        'error'
+      )
+    }
+  })
+}
+
 }
